/ppbox/branches/1.2.0 /src/ppbox/dac 修正数值转换精度范围丢失编译警告
git-svn-id: http://svn01/repositories/r01/ppbox/branches/1.2.0@5065 b3a1389e-2159-5c43-9038-463bf877d0d9

+++ asf/AsfMuxer.h
@@ -36,7 +36,7 @@ namespace ppbox
                 ppbox::demux::Sample & tag);
 
             boost::system::error_code seek(
-                boost::uint32_t & time,
+                boost::uint64_t & time,
                 boost::system::error_code & ec);
 
         private:
